sarah-ai-companion/multimodal_engine/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Image Generation
torch==2.1.0
torchvision==0.16.0
diffusers==0.24.0
transformers==4.36.0
accelerate==0.25.0
safetensors==0.4.1
xformers==0.0.23
huggingface-hub==0.19.4

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2

# LoRA Training
peft==0.7.1
datasets==2.15.0
bitsandbytes==0.41.3

# TTS Integration
httpx==0.25.2
aiohttp==3.9.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# File Management
aiofiles==23.2.1

# JSON Processing
orjson==3.9.10

# Logging
python-json-logger==2.0.7

# Background Tasks
celery==5.3.4
redis==5.0.1

# Type Hints
typing-extensions==4.8.0
